Move ProjectCarrousel to shared folder
